条件式の修正

+++ app/controllers/items_controller.rb
@@ -36,8 +36,10 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    if @item.destroy
-      redirect_to root_path
+    if current_user.id == @item.user.id
+      if @item.destroy
+        redirect_to root_path
+      end
     end
   end
